% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_number.R
\name{power_number}
\alias{power_number}
\title{Calculate the nth power of a number}
\usage{
power_number(x, n)
}
\arguments{
\item{x}{Numeric. Number to take the power of.}

\item{n}{Numeric. Power to take}
}
\value{
\code{x^n}
}
\description{
Calculate the nth power of a number
}
\examples{

power_number(5, 2)
power_number(1.3, 0.56)

}
